Optimized protocol module code.
When the viewer loads a file it needs to list all the files in the containing
directory in order to determine the index of that filename within the directory.
Previously ProtocolDirectoryListing was getting the file list and getting the
filenames for all the entries. Recent changes I made to my custom protocol made
getting the filename for a given path a relatively expensive operation,
requiring a full request cycle to the server to determine the file's file type.
This modification moves the index determination logic into the external protocol
module, which is able to find the file in the file listing via cheaper, local
operations.